flask_app/requirements.txt
celery[redis]
flask
pymongo
celery
redis
python-dotenv
pandas
openpyxl